fix: repair tests
